text_to_sql_llm/requirement.txt.txt
streamlit
google-generativeai
python-dotenv
#langchain
#PyPDF2
#chromadb
#faiss-cpu
#pdf2image
#sqlite3